requirements.txt
openai==1.45.0
langchain==0.3.0
langchain-community==0.3.0
langchain_openai==0.3.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.145.0
pydantic==2.9.2
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.6
langchain-core==0.3.5
